+++ src/main/java/apiquality/sonar/asyncapi/checks/RulesLists.java
@@ -33,8 +33,8 @@ public final class RulesLists {
             AAR035MessageTitleCheck.class,
             AAR036BadDescriptionCheck.class,
             AAR037BindingVersionCheck.class,
-            AAR038InfoTitleRequiredCheck.class,
-            AAR042MessageIdentifierCheck.class
+            AAR042MessageIdentifierCheck.class,
+            AAR050InfoTitleRequiredCheck.class
         );
     }